// shared/model/catalogos/puesto.ts
export interface PuestoI {    
    idDepto?: string;
    idPuesto?: string;
    puesto: string;
    responsabilidad?: string;
    departamento?: string;
    id_departamento?: string;
    disp_cambio?: boolean;
    disp_viajar?: boolean;
    formacion?: string;
    habilidades?: string;        
    requisitos?: string;    
    remuneracion_de?: number;
    remuneracion_hasta?: number;  
    deshabilitar?: boolean;
}